@@ -2,31 +2,29 @@
   "name": "flight_plan",
   "private": true,
   "dependencies": {
-    "@githubprimer/octicons-react": "^8.1.0",
-    "@reach/dialog": "^0.13.2",
-    "@reach/menu-button": "^0.13.2",
-    "@reach/portal": "^0.13.2",
-    "@reach/router": "^1.3.4",
-    "@reach/tooltip": "^0.13.2",
-    "@reduxjs/toolkit": "^1.5.0",
-    "@tailwindcss/typography": "^0.4.0",
-    "actioncable": "5.2.4-5",
-    "autoprefixer": "^10.2.1",
-    "classnames": "^2.2.6",
-    "downshift": "^6.1.0",
-    "font-color-contrast": "^1.0.3",
-    "normalizr": "^3.3.0",
-    "postcss": "^8.2.4",
-    "rails-ujs": "5.2.4-5",
-    "react": "^16.14.0",
-    "react-beautiful-dnd": "^13.0.0",
-    "react-dom": "16.14.0",
-    "react-redux": "^7.2.2",
-    "react-scrollbar-size": "^4.0.0",
-    "react-toastify": "^7.0.3",
-    "redux": "^4.0.5",
-    "showdown": "^1.9.1",
-    "tailwindcss": "^2.0.2"
+    "@floating-ui/react": "^0.22.2",
+    "@gatsbyjs/reach-router": "^2.0.1",
+    "@headlessui/react": "^1.7.13",
+    "@primer/octicons-react": "^18.3.0",
+    "@reduxjs/toolkit": "^1.9.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "actioncable": "^5.2.8-1",
+    "autoprefixer": "^10.4.14",
+    "classnames": "^2.3.2",
+    "downshift": "^7.6.0",
+    "font-color-contrast": "^11.1.0",
+    "normalizr": "^3.6.2",
+    "postcss": "^8.4.21",
+    "rails-ujs": "^5.2.8-1",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-scrollbar-size": "^5.0.0",
+    "react-toastify": "^9.1.2",
+    "redux": "^4.2.1",
+    "showdown": "^2.1.0",
+    "tailwindcss": "^3.3.0"
   },
   "engines": {
     "node": "18.14.0",
@@ -42,8 +40,8 @@
     }
   },
   "devDependencies": {
-    "@vitejs/plugin-react-swc": "^3.0.1",
-    "vite": "^4.0.0",
-    "vite-plugin-ruby": "^3.1.2"
+    "@vitejs/plugin-react-swc": "^3.2.0",
+    "vite": "^4.2.0",
+    "vite-plugin-ruby": "^3.2.0"
   }
 }
