@@ -1,16 +1,8 @@
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import Root from "../demos";
 
-const rootRender = () => {
-    ReactDOM.render(
-        <Root />,
-        document.querySelector("#root")
-    );
-};
+const rootRender = () =>
+    createRoot(document.querySelector("#root")).render(<Root />);
 
 document.addEventListener("DOMContentLoaded", rootRender);
-
-if (module.hot) {
-    module.hot.accept("../demos", rootRender);
-}
