@@ -24,7 +24,7 @@ const Entry = ({ author = "ghost", body, timestamp, action = "commented" }) => (
     <div className="flex mb-4">
         <Avatar username={author} />
 
-        <div className="ml-3 flex-grow" style={{ width: "calc(100% - 3rem)" }}>
+        <div className="ml-3 grow w-[calc(100%-3rem)]">
             <div className="border border-gray-300 rounded-t bg-gray-100 px-3 py-2 font-bold text-sm">
                 <a
                     href={`https://github.com/${author}`}
