@@ -1,9 +1,9 @@
 import { Fragment } from "react";
 import { connect } from "react-redux";
-import { Link } from "@reach/router";
+import { Link } from "@gatsbyjs/reach-router";
 import { denormalize, schema } from "normalizr";
 import classNames from "classnames";
-import Octicon, { Gear } from "@githubprimer/octicons-react";
+import { GearIcon } from "@primer/octicons-react";
 
 import Label, { Milestone } from "./Label";
 import PullRequest from "./PullRequest";
@@ -57,8 +57,8 @@ const SidebarEntry = ({ title, children, url }) => (
                     to={url}
                     className="w-full flex items-center justify-start hover:text-blue-500 focus:text-blue-500 focus:outline-none"
                 >
-                    <span className="flex-grow text-left">{title}</span>
-                    <Octicon icon={Gear} />
+                    <span className="grow text-left">{title}</span>
+                    <GearIcon />
                 </Link>
             ) : title}
         </h2>
