@@ -1,4 +1,4 @@
-const features = window.flightPlanConfig.features || {};
+const features = () => window.flightPlanConfig.features || {};
 
-export const isFeatureEnabled = name => !!features[name];
+export const isFeatureEnabled = name => !!features()[name];
 export const isFeatureDisabled = name => !isFeatureEnabled(name);
